requirements.txt
torch==1.2.0+cu92
torchvision==0.4.0+cu92
numpy==1.18.0
Pillow==6.2.1
dlib==19.4.0
tqdm==4.41.0